+++ CHANGELOG.md
@@ -18,7 +18,9 @@ Security in case of vulnerabilities.
 
 ### Added
 - Added `Complete a notification of changes to funded high needs places form` task for conversion projects.
+- Added 'Caching to redis api call'
 - Added transfer creation end point for prepare to complete
+
 
 ### Changed
 
@@ -28,6 +30,8 @@ Security in case of vulnerabilities.
 - minor text correction for confirm dao revocation page
 - Fixed task notes formatting.
 - Replaced html extension method with note html tag to have consistent approach for all notes/reasoning  formatting.
+
+- Fixed `LocalAuthority` HTTP Post method by removing local authority id.
 
 ### Security